undid comment and "fix" re: Jacobian of Poisson
Lambda in Poisson is just a parameterization convention. Since q is
measured wrt eta ~ Normal, we should measure p wrt the same. But we can
parameterize the distribution however we want, so long as we put an
appropriate measure on the parameters (= pdf + differentials). No need
to include a Jacobian when using a nonstandard parameterization of
Poisson, unless the integral is wrt the canonical parameter.